chore(client/ts): port setup.js entrypoint to TS
